Fix Issues #2, #4 and #5 (#6)
* Fixed the issues on the imports.
* Modified the download function to enable parallel exploration of the generator and provided a minimal example on how to perform parallel downloading of a dataset.
* Split of dataset.py into sub files (remote_dataset.py and local_dataset.py) and introduced two support scripts upload_manager.py
* Added documentation on the downloading functions
