+++ darwin/__init__.py
@@ -1,4 +1,4 @@
 import darwin.exceptions
-from darwin.client import Client
-from darwin.dataset import Dataset, LocalDataset
-from darwin.team import Team
+
+from .client import Client
+from .team import Team
