+++ darwin/cli_functions.py
@@ -8,7 +8,6 @@ from tqdm import tqdm
 
 from darwin.client import Client
 from darwin.config import Config
-from darwin.dataset import SUPPORTED_IMAGE_EXTENSIONS, SUPPORTED_VIDEO_EXTENSIONS
 from darwin.exceptions import (
     InvalidLogin,
     MissingConfig,
@@ -18,6 +17,7 @@ from darwin.exceptions import (
     ValidationError,
 )
 from darwin.table import Table
+from darwin.utils import SUPPORTED_IMAGE_EXTENSIONS, SUPPORTED_VIDEO_EXTENSIONS
 
 
 def error(message):
@@ -50,10 +50,10 @@ def authenticate(email: str, password: str, projects_dir: str) -> Config:
     config_path.parent.mkdir(exist_ok=True)
 
     default_config = {
-        "token": client._token,
-        "refresh_token": client._refresh_token,
-        "api_endpoint": client._url,
-        "base_url": client._base_url,
+        "token": client.token,
+        "refresh_token": client.refresh_token,
+        "api_endpoint": client.url,
+        "base_url": client.base_url,
         "projects_dir": projects_dir,
     }
 
@@ -92,20 +92,9 @@ def set_team(team_slug: str):
     config.write("refresh_token", client._refresh_token)
 
 
-def continue_request():
-    """Asks for explicit approval from the user. """
-    approval = input("Do you want to continue? [Y/n] ")
-    if approval not in ["Y", "y", ""]:
-        return False
-    return True
-
-
 def secure_continue_request():
-    """Asks for explicit approval from the user. """
-    approval = input("Do you want to continue? [y/N] ")
-    if approval not in ["Y", "y"]:
-        return False
-    return True
+    """Asks for explicit approval from the user. Empty string not accepted"""
+    return input("Do you want to continue? [y/N] ") in ["Y", "y"]
 
 
 def create_dataset(dataset_name: str):
@@ -146,7 +135,9 @@ def path(project_slug: str) -> Path:
         local_dataset = client.get_local_dataset(slug=project_slug)
     except NotFound:
         error(
-            f"Project '{project_slug}' does not exist locally. Use 'darwin remote' to see all the available projects, and 'darwin pull' to pull them."
+            f"Project '{project_slug}' does not exist locally. "
+            f"Use 'darwin remote' to see all the available projects, "
+            f"and 'darwin pull' to pull them."
         )
     return local_dataset.project_path
 
@@ -162,20 +153,21 @@ def url(project_slug: str) -> Path:
 
 
 def pull_project(project_slug: str):
-    """Downloads a rermote project (images and annotations) in the projects directory. """
+    """Downloads a remote project (images and annotations) in the projects directory. """
     client = load_client()
     try:
         dataset = client.get_remote_dataset(slug=project_slug)
     except NotFound:
         error(
-            f"project '{project_slug}' does not exist at {client._url}. Use 'darwin remote' to list all the remote projects."
+            f"project '{project_slug}' does not exist at {client.url}. "
+            f"Use 'darwin remote' to list all the remote projects."
         )
     except Unauthenticated:
         error(f"please re-authenticate")
     print(f"Pulling project {project_slug}:latest")
     progress, count = dataset.pull()
-    for _ in tqdm(progress(), total=count, desc="Downloading"):
-        pass
+    for f in tqdm(progress(), total=count, desc="Downloading"):
+        f()
 
     return dataset.local()
